add the Step abstraction
A Step is a function to call, and a pointer to pass it.

- add step.h defining the external API
- add step.c providing the implementation
- call step_post, step_bist, and step_bench from main.c

Add an "ASSERT_EQ_integer" macro to represent a test
that requires two expressions to evaluate to the same
integer value; on failure, print expected and observed
values, and the expressions in string form that were
passed to the macro.

Also:
- prune the -T list from the .indent.pro file