add SigTrace and friends
I needed a way to display timing diagrams for the simulated
signals, and have decided to use the "schemdraw" package,
which uses WaveDrom. I define the plots using JSON strings
dynamically created from logged trace data; there is the notion
of a trace session, from which multiple plots can be generated.

The display package accepts TeX (or TeX-like?) strings for the
labels, so a few adjustments to edge names are in order.

While getting it working, I added signal tracing to the i8224
simulation, then generated a nice ORG document that includes
the diagrams inline (toggle inline images "on" in the ORG buffer).

Then I added tracing to the i8228 which has a larger number
of less interesting plots. However, having the plots available
made it easy to cross-check that all is well.