add Addr and Data for Timing Plots
Derive Addr and Data types from Edge, which allow tracking of
values of 16-bit (Addr) and 8-bit (Data) entities. This starts with
the Address and Data busses but can and will be used for the
various 8-bit and 16-bit latched data that might be useful to
include on a timing diagram.

Expand sigtrace to handle Addr and Data items.

Decline to include signals in plots that do not move.

Update existing plots to include ADDR and DATA as apprioriate.

Use addr_z and data_z at appropriate times so that timing plots
show when the busses return to high impedence.

Also:
- GIVE UP ON OCTAL. I just can't think in Octal any more.
- EDGE_INIT and edge_init expect an initial value.
- make address handling logic more explicit in decoder