add IDAL internal block
This models the Incrementer-Decrementer Address Latch
which is architectural state of the 8080 but is not visible
to the programmer.

I infer from behavior that it latches the 16-bit data for the
address bus output, and presents the incremented data
back to the address bank to be stored, since the incremented
data is noted as being stored in T2 but the PC address must
stay valid well into T3.