Auto-drop SHADOW
SHADOW is now dropped by the Decoder automatically
if the properly decoded edge matches the shadow edge.

This means that we need to present each HEX file to each ROM
at each address the ROM maps, which means that we should not
erase the ROM before loading the file. This appears twice in the
current i8080_asm_init code, which may need to be expanded to
load all ROMs in the future.

(Do we really need more than 2 KiB of ROM?)

Use an edge-trigger to re-establish SHADOW if RESET rises.

Also:
- rom[0] is at the top of memory, rom[1] below it, and so on.
- RESET forces SYNC DBIN WAIT state to zero but does not trigger callbacks;
  in particular we want to avoid triggering DBIN falling edge with ADDR set to "z"
- turn "abort if we hit an unimp instruction" back on and fix the printed PC.
- decoder test did not init its ADDR storage