add Bring and Cdev
Bring: thread-safe ring of bytes with a single thread
producing and a single thread consuming, with fixed
capacity of 255 bytes.

Cdev: Character Device that appears in the simulated
device space as a read/write data register, a read-only
status register, and a write-only control register; the
back-end of cdev is an INET socket connecting to an
ancient "tcp-tty" server (details not interesting for this
project).