import "sio" utility and friends
Recent events on my dev machine have convinced me to include
the device side of my SIO simulation in the project.

This in fact was what drove the "inc" rework just completed:
this program includes <pthread.h> which includes <sched.h>
which was resolving to the local inc/sched.h rather than the
system <sched.h> that it actually needs.

The build system is now set up to expect that each ".c" file
found in utl/ is a utility to be built and installed in bin/

- sio: a TCP server simulating a connected serial device
- sio-loop.sh: call sio repeatedly in this terminal window,
  adjusting TTY modes (with stty) as appropriate.
- sio-crt.sh: use sio to provide a Video Terminal
- sio-tty.sh: use sio to provide a Printing Terminal
- sio-lot.sh: use sio to provide a Line Printer