add bdev-format.c and fmt.sh
The bdev-format.c file is where I tested that "mmap" still works
sanely (and with the same crazy warts we love) on Ubuntu Linux
inside Windows WSL-2. The public purpose of this program is to
create the files that BDEV uses to persist its data fro run to run,
to format that data appropriately, and to pre-load boot images
onto the start of the drives as needed.

fmt.sh establishes the drives as:
- "A" is an IBM Single Density Disk that should boot CP/M
- "B" is an IBM Single Density Disk that should boot FORTH
- "C" is a hard drive storing about 2 MB of data
- "D" is a hard drive storing about 8 MB of data

Also:
- in CBIOS: move fixed tables to below the code segment.
- tell "indent" that HexStoreFn is a type, and reindent.